--- a/.changeset/moody-comics-exist.md
+++ b/.changeset/moody-comics-exist.md
@@ -1,0 +1,5 @@
+---
+"react-router": minor
+---
+
+Optimize Lazy Route Discovery path discovery to favor a single `querySelectorAll` call at the `body` level instead of many calls at the sub-tree level