--- a/contributors.yml
+++ b/contributors.yml
@@ -76,6 +76,7 @@
 - david-crespo
 - dcblair
 - decadentsavant
+- developit
 - dgrijuela
 - DigitalNaut
 - dmitrytarassov