New PR for https://github.com/remix-run/react-router/pull/12542 pointed to `dev`

In a real-world website I am testing this against, this reduces the fog-of-war's blocking JS during SPA transitions from 100ms to 1.2ms.

|Before|After|
|---|---|
| <img width="470" alt="Screenshot 2024-12-13 at 12 05 06 PM" src="https://github.com/user-attachments/assets/b4c59150-d1ce-406c-9c4d-6c794354c32c" /> | <img width="441" alt="Screenshot 2024-12-13 at 12 05 36 PM" src="https://github.com/user-attachments/assets/fcb38b8f-dd7f-475a-bd8d-efad52e82202" /> |